vim-patch:8.1.2104: the normal.c file is too big
Problem:    The normal.c file is too big.
Solution:   Move do_pending_operator() to ops.c. (Yegappan Lakshmanan,
            closes vim/vim#4999).
https://github.com/vim/vim/commit/792cf5e1bec04c6d6d70cfbb9ef24c798b469731
